Control.LVish.DeepFrz.Internal

Description

This module is not Safe Haskell, but it must be used to create new LVar types.

Documentation

class DeepFrz a where

DeepFreezing is a type-level (guaranteed O(1) time complexity) operation. It marks an LVar and its contents (recursively) as frozen. DeepFreezing is not an action that can be taken directly by the user, however. Rather, it is the final step in a runParThenFreeze invocation.

Associated Types

type FrzType a :: *

This type function is public. It maps pre-frozen types to frozen ones. It should be idempotent.

Methods

frz :: a -> FrzType a

Private: not exported to the end user.

data NonFrzn

This exists only for the purpose of being a type which is not equal to Frzn. One could just as well have used (), but this is more descriptive.

data Frzn

An uninhabited type that signals that an LVar has been frozen. LVars should use this in place of their s parameter.

data Trvrsbl

An uninhabited type that signals that an LVar is not only frozen, but it may be traversed in whatever order its internal representation dictates.