Data.LVar.AddRemoveSet

Description

This module provides sets that allow both addition and removal of elements. This is possible because, under the hood, it's represented with two monotonically growing sets, one for additions and one for removals. It is inspired by 2P-Sets from the literature on conflict-free replicated data types.

Documentation

data AddRemoveSet s a

The set datatype.

newEmptySet :: Ord a => Par e s (AddRemoveSet s a)

Create a new, empty AddRemoveSet.

newSet :: Ord a => Set a -> Par e s (AddRemoveSet s a)

Create a new AddRemoveSet populated with initial elements.

newFromList :: Ord a => [a] -> Par e s (AddRemoveSet s a)

A simple convenience function. Create a new ISet drawing initial elements from an existing list.

insert :: (HasPut e, Ord a) => a -> AddRemoveSet s a -> Par e s ()

Put a single element in the set. (WHNF) Strict in the element being put in the set.

waitAddedElem :: (HasGet e, Ord a) => a -> AddRemoveSet s a -> Par e s ()

Wait for the set to contain a specified element.

waitAddedSize :: HasGet e => Int -> AddRemoveSet s a -> Par e s ()

Wait on the size of the set of added elements.

remove :: (HasPut e, Ord a) => a -> AddRemoveSet s a -> Par e s ()

Remove a single element from the set.

waitRemovedElem :: (HasGet e, Ord a) => a -> AddRemoveSet s a -> Par e s ()

Wait for a single element to be removed from the set.

waitRemovedSize :: HasGet e => Int -> AddRemoveSet s a -> Par e s ()

Wait on the size of the set of removed elements.

freezeSet :: (Ord a, HasFreeze e) => AddRemoveSet s a -> Par e s (Set a)

Get the exact contents of the set. As with any quasi-deterministic operation, using freezeSet may cause your program to exhibit a limited form of nondeterminism: it will never return the wrong answer, but it may include synchronization bugs that can (nondeterministically) cause exceptions.