Data.LVar.IStructure

Description

An I-Structure, also known as an array of IVars, implemented using a boxed vector.

Documentation

data IStructure s a

An I-Structure, also known as an array of IVars.

Basic operations

newIStructure :: Int -> Par e s (IStructure s elt)

Create a new, empty, monotonically growing IStructure of a given size. All entries start off as zero, which must be "bottom".

newIStructureWithCallback :: Int -> (Int -> elt -> Par e s ()) -> Par e s (IStructure s elt)

Register handlers on each internal IVar as it is created. This operation should be more efficient than newIStructure followed by forEachHP.

put :: (NFData elt, Eq elt, HasPut e) => IStructure s elt -> Int -> elt -> Par e s ()

Put a single element in the IStructure at a given index. This variant is deeply strict (NFData).

put_ :: (Eq elt, HasPut e) => IStructure s elt -> Int -> elt -> Par e s ()

Put a single element in the IStructure at a given index. That index must be previously empty. (WHNF) Strict in the element being put in the set.

get :: (Eq elt, HasGet e) => IStructure s elt -> Int -> Par e s elt

Wait for the indexed entry to contain a value, and return that value.

getLength :: IStructure s a -> Par e s Int

Retrieve the number of slots in the IStructure.

Iteration and callbacks

forEachHP

Arguments

:: Maybe HandlerPool
pool to enroll in, if any

-> IStructure s a
IStructure to listen to

-> (Int -> a -> Par e s ())
callback

-> Par e s ()

Add an (asynchronous) callback that listens for all new elements added to the IStructure, optionally enrolled in a handler pool.

Freezing

freezeIStructure :: HasFreeze e => IStructure s a -> Par e s (Vector (Maybe a))

O(N) complexity, unfortunately. This implementation of IStructures requires freezing each of the individual IVars stored in the array.