Data.LVar.PNCounter

Description

This module provides a PN-Counter, a counter that allows both increment and decrement operations. This is possible because, under the hood, it's represented with two monotonically growing counters, one for increments and one for decrements. The name PN-Counter comes from the literature on conflict-free replicated data types.

Documentation

data PNCounter s

The counter datatype.

newCounter :: Par e s (PNCounter s)

Create a new PNCounter set to zero.

newCounterWithValue :: Int -> Par e s (PNCounter s)

Create a new PNCounter with the specified initial value.

increment :: HasBump e => PNCounter s -> Par e s ()

Increment the PNCounter.

waitForIncrements :: HasGet e => Int -> PNCounter s -> Par e s ()

Wait for the number of increments to reach a given number.

decrement :: HasBump e => PNCounter s -> Par e s ()

Decrement the PNCounter.

waitForDecrements :: HasGet e => Int -> PNCounter s -> Par e s ()

Wait for the number of decrements to reach a given number.

freezeCounter :: HasFreeze e => PNCounter s -> Par e s Int

Get the exact contents of the counter. As with any quasi-deterministic operation, using freezeCounter may cause your program to exhibit a limited form of nondeterminism: it will never return the wrong answer, but it may include synchronization bugs that can (nondeterministically) cause exceptions.