Data.LVar.PureMap.Unsafe

Unsafe operations:

unsafePeekKey :: Ord k => k -> IMap k s v -> Par e s (Maybe v)

An unsafe, nonblocking version of getKey. This reveals whether

These are here only to reexport downstream:

newtype IMap k s v

The map datatype itself. Like all other LVars, it has an s parameter (think STRef) in addition to the a parameter that describes the type of elements in the set.

Performance note: There is only one mutable location in this implementation. Thus it is not a scalable implementation.

Constructors

IMap (LVar s (IORef (Map k v)) (k, v))

forEachHP

Arguments

:: Maybe HandlerPool
optional pool to enroll in

-> IMap k s v
Map to listen to

-> (k -> v -> Par e s ())
callback

-> Par e s ()

Add an (asynchronous) callback that listens for all new key/value pairs added to the map, optionally enrolled in a handler pool.